LayoutTests/accessibility/aria-labeled-with-hidden-node.html

<!DOCTYPE html>
<html>
<title id="title">TITLE</title>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>

<button id="button" aria-labelledby="title">Go</button>


<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that if try to use a non-rendered node as your labelledby element, it won't crash and it will work.");
    
    if (window.accessibilityController) {
       var button = accessibilityController.accessibleElementById("button");
       if (accessibilityController.platformName == "atk")
           shouldBe("button.title", "'AXTitle: TITLE'");
       else
           shouldBe("button.description", "'AXDescription: TITLE'");
    }
</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
